requirements.txt
openai
langchain
python-gitlab
atlassian-python-api
python-dotenv
